util: Avoid std::function overhead in ScopeGuard
The type of the callable is usually just a lambda or a function pointer
so there is no need to rely on std::function's type erasure.

While the cost of using std::function is probably negligible, it still
causes some unnecessary overhead that can be avoided by making
ScopeGuard a templated class. Usage is still as easy thanks to template
class argument deduction in C++17.

See https://godbolt.org/z/KcoPni for a comparison between
a ScopeGuard that uses std::function and one that doesn't